AtCoder/ABC366/C/main_test.go
package main

import (
	"bytes"
	"io"
	"strings"
	"testing"
)

func Test_run(t *testing.T) {
	type args struct {
		stdin io.Reader
	}
	tests := []struct {
		name    string
		args    args
		wantOut string
	}{
		{name: "1", args: args{stdin: strings.NewReader(`8
1 3
1 1
1 4
3
2 1
3
1 5
3`)}, wantOut: `3
2
3
`},
		{name: "2", args: args{stdin: strings.NewReader(`8
1 2
1 2
3
2 2
1 4
1 4
2 2
3`)}, wantOut: `1
1
`},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			out := &bytes.Buffer{}
			run(tt.args.stdin, out)
			if gotOut := out.String(); gotOut != tt.wantOut {
				t.Errorf("run() = %v, want %v", gotOut, tt.wantOut)
			}
		})
	}
}
